+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## v1.1.0
 
+- feat(protocol): added unix socket supports.
 - feat(protocol): accepts socket factory function to produce sockets.
 - feat(encoder): accepts string as data to send.
 - feat(connector): added `IConnection.writable` property.